# services/reference-service/src/main/resources/openapi/reference-api.yaml
...
openapi: 3.1.0
info:
  title: CardDemo Reference Service API
  version: 1.0.0
  summary: >-
    Reference data of the migrated CardDemo credit-card management application:
    transaction types and their categories, disclosure-group interest rates, the
    three United States address lookups, and date evaluation.
  description: >-
    The reference context owns the data every other context reads and none of
    them writes. It publishes nineteen operations, across thirteen paths, over
    six tables.


    Transaction types and transaction categories are maintained here, and the
    relationship between them is the one referential rule the baseline states
    anywhere in this domain: a category references its type, and the type cannot
    be deleted while a category still references it. That refusal is a
    first-class user outcome in the baseline rather than a fault, and it is
    reported here as 409 for the reason recorded on the Conflict response.


    Disclosure-group rows carry the interest rate the nightly interest
    calculation consumes. That rate is an arithmetic operand, not a display
    value, so it is transported as a JSON string for the reason recorded on the
    InterestRate schema, and the lookup operation documents the DEFAULT-group
    fallback the baseline performs when an account's own group has no matching
    row.


    The three address lookups -- phone area codes, state codes and state and
...
    operation that writes them.


...


...
    one field.
...

...
servers:
...
security:
  - bearerAuth: []

# WHY : Refactoring Rationale: the authority each operation requires is declared
#       as a machine-readable field on the operation itself, and not only in
#       prose and tag membership. Prose and tag membership are not checkable: a
#       reviewer can read that a write is administrative, and a generator or a
#       gateway cannot, so the restriction would be indistinguishable from any
#       other authenticated route. Every operation below carries
#       x-required-authority.
# WHY : Assumptions: the value on each operation is not chosen here. It is
#       derived from what SecurityConfig.java already enforces for this service:
#       POST, PUT, PATCH and DELETE on every path each require
#       JwtRoleConverter.ADMIN_AUTHORITY, at its lines 99 to 106, and every
#       remaining request needs only authentication, at line 107. So every write
#       below carries carddemo-admin and every READ carries the third value
#       authenticated, and the document describes the filter chain rather than a
#       second opinion about it.
# WHY : Refactoring Rationale: an earlier revision carried carddemo-user on the
#       reads, which stated something the filter chain does not enforce and could
#       not be enforced without a behavioural regression. The shared converter
#       grants exactly the groups a token presents and builds NO hierarchy -- its
#       conversion loop admits a group only if it is one of the two recognised
#       names -- so a token carrying carddemo-admin alone receives ADMIN_AUTHORITY
#       and nothing else. A gateway or generator enforcing carddemo-user on the
#       reads would therefore refuse every administrator, and the baseline refuses
#       an administrator nothing: app/cbl/COSGN00C.cbl sends type 'A' to the
#       administrative menu at its line 245 and every function is reachable from
#       there. The honest declaration is the one the chain makes.
# WHY : Alternatives Considered: granting USER_AUTHORITY to an administrator as
#       well, so that carddemo-user could be declared and enforced on the reads.
#       Rejected because the grant would have to be made in the shared converter,
#       where it would change the authority set every one of the nine contexts
#       sees, to buy nothing this service needs: a read that any authenticated
#       caller may perform is described exactly by saying so. The two GROUP values
#       remain exactly the two the shared converter recognises, ADMIN_AUTHORITY
#       and USER_AUTHORITY at
#       services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java
#       lines 250 and 260, which are the migrated form of the two SEC-USR-TYPE
#       values 'A' and 'U'; authenticated names a REQUIREMENT rather than a group,
#       and the model below distinguishes the two.
...
paths:
...
  /api/v1/reference/transaction-categories:

    description: >-
      The collection of transaction categories, each belonging to exactly one
      transaction type. A category is identified by the pair of its type code and
      its own four-digit code, which is why the single-category path below
      carries two segments rather than one.

    get:
      tags:
        - reference-data
      operationId: listTransactionCategories
      x-required-authority: authenticated
      summary: List transaction categories by key, one page at a time
      description: >-
        Returns one page of transaction categories ordered by type code and then
        by category code, optionally narrowed to a single type. That ordering is
        the baseline's own: the composite key of reference.transaction_categories
        is declared in that column order by db/migration/V1__reference.sql, and
        the baseline index at
        app/app-transaction-type-db2/ddl/XTRNTYCAT.ddl declares the same two
        columns ascending in the same sequence.


        Paging follows the identical keyset discipline as the transaction-type
        list, over the composite key rather than a single column, so a cursor
        from one collection is not interchangeable with a cursor from the other.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PagingDirection'
        - $ref: '#/components/parameters/TypeCodeFilter'
        - $ref: '#/components/parameters/DescriptionFilter'
      responses:
        '200':
          description: >-
            One page of transaction categories with the two cursor positions and
            the forward availability flag.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategoryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - admin-reference-data
      operationId: createTransactionCategory
      x-required-authority: carddemo-admin
      summary: Create a transaction category under an existing type
      description: >-
        Creates one category beneath an existing transaction type. The type must
        already exist: the foreign key declared by
        db/migration/V1__reference.sql refers to reference.transaction_types, so
        a category naming an absent type is refused with 409 rather than stored
        as an orphan. This is the same rule that refuses the parent's delete,
        seen from the other side, and it reaches this status through the same
        integrity branch, so it carries the same referential sentence -- as does a
        duplicate pair of codes, for the reason registered as
        D-REFERENCE-INTEGRITY-SENTENCE in
        docs/architecture/cobol-to-service-traceability.md.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      requestBody:
        required: true
        description: >-
          The type code, the category code and the description of the category
          to create. All three are required, because all three columns are
          declared NOT NULL.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryCreateRequest'
      responses:
        '201':
          description: >-
            The category was created, and the stored representation is returned.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
            Location:
              description: >-
                The path of the created category, formed from the two codes
                supplied in the request body.
              required: true
              schema:
                type: string
                examples:
                  - /api/v1/reference/transaction-categories/01/0006
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/transaction-categories/{typeCd}/{catCd}:

    description: >-
      One transaction category, addressed by the pair of codes that forms its
      composite key. Both segments are present because neither alone identifies
      a row: the same four-digit category code recurs under different types, and
      the baseline builds its key by positional concatenation of the two, which
      is why the seed extract app/data/ASCII/trancatg.txt stores the pair as the
      single six-character value "010001".

    parameters:
      - $ref: '#/components/parameters/TypeCdPath'
      - $ref: '#/components/parameters/CatCdPath'

    get:
      tags:
        - reference-data
      operationId: getTransactionCategory
      x-required-authority: authenticated
      summary: Read one transaction category
      description: >-
        Returns the category stored under the supplied pair of codes, together
        with the version token a subsequent replace must carry.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: >-
            The stored transaction category, including its version token.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - admin-reference-data
      operationId: replaceTransactionCategory
      x-required-authority: carddemo-admin
      summary: Replace the description of one transaction category
      description: >-
        Replaces the description of an existing category. Neither code is
        replaceable, for the same reason the type code is not: together they are
        the primary key. The request must carry the version token read with the
        record, and a stored row that has moved on since that read is reported as
        409 with the baseline's own wording.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      requestBody:
        required: true
        description: >-
          The replacement description together with the version token read with
          the record.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryReplaceRequest'
      responses:
        '200':
          description: >-
            The category was replaced, and the stored representation is returned
            carrying its new version token.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - admin-reference-data
      operationId: deleteTransactionCategory
      x-required-authority: carddemo-admin
      summary: Delete one transaction category
      description: >-
        Deletes the category stored under the supplied pair of codes. Nothing
        references a category, so this delete is not refused on referential
        grounds the way a type's delete can be: no table in the reference schema
        declares a foreign key to reference.transaction_categories, and
        db/migration/V1__reference.sql records why -- the disclosure-group table
        carries the same type and category columns but the baseline states no
        relationship between them, so none is asserted.


        Deleting a category that no longer exists is reported as 404 rather than
        as success, so that a caller can distinguish having removed the row from
        never having found it.


        This operation can still answer 409, and the reason is narrower than for a
        type's delete: the referential condition cannot arise here, but the
        pessimistic-lock condition can arise on any write, so the shared Conflict
        response is declared and its lock sentence is the one this operation can
        return.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '204':
          description: >-
            The category was deleted. No body is returned.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A JSON Web Token issued by the configured identity provider and validated
        as a resource-server token. Its cognito:groups claim is converted to
        authorities by
        services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java,
        where the baseline user type 'A' becomes carddemo-admin and 'U' becomes
        carddemo-user. Presenting no usable token yields 401; presenting a valid
        token that lacks the authority an operation requires yields 403.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: >-
        The paging position to continue from: copy the lastKey member of the page
        in hand to advance past it, or the firstKey member to retreat from it, and
        state the matching direction. Omit it, and the direction, to obtain the
        opening page. The value is opaque and must be echoed exactly as received;
        it is not a type code, a category code or a row ordinal, and it cannot be
        constructed, parsed or compared by a client. A cursor that cannot be
        opened, or one replayed against a different filter, a different collection
        or the direction it was not minted for, is reported as 400 rather than
        answered with the wrong page.
      schema:
        $ref: '#/components/schemas/CursorToken'
    PagingDirection:
      name: direction
      in: query
      required: false
      description: >-
        Which way to move from the supplied cursor. It is meaningful only
        alongside a cursor, and the two are sent together or not at all. Reading
        next takes rows from the cursor's position forward in ascending key order;
        reading previous takes rows before it in descending key order and presents
        them ascending, which is the migrated form of the backward cursor at
        app/app-transaction-type-db2/cbl/COTRTLIC.cbl lines 354 to 368. The two
        differ in whether the cursor's own position is included, as recorded on
        the cursor parameter.
      schema:
        $ref: '#/components/schemas/PageDirection'
    TypeCodeFilter:
      name: typeCode
      in: query
      required: false
      description: >-
        Narrows the page to the single transaction type whose code equals this
        value. This is the migrated form of the optional equality predicate the
        baseline carries in both cursors, at
        app/app-transaction-type-db2/cbl/COTRTLIC.cbl lines 344 to 345 going
        forward and 360 to 361 going back, each gated on an edit flag whose
        unset state passes every row through -- the pass-through arms at lines 346
        and 362. Omitting this parameter is that unset state, and so is sending
        '00': the baseline collapses low values, spaces and zeros into one
        filter-blank state at lines 1101 to 1106. That is why this parameter takes
        TransactionTypeCodeFilter rather than the stored-code schema, which refuses
        '00' outright. A value that is not two digits is reported as 400 carrying the
        baseline's own wording, which is recorded on the ReferenceMessageCatalogue
        schema.
      schema:
        $ref: '#/components/schemas/TransactionTypeCodeFilter'
    DescriptionFilter:
      name: description
      in: query
      required: false
      description: >-
        Narrows the page to rows whose description CONTAINS this text. The baseline
        semantic is a SQL LIKE, at app/app-transaction-type-db2/cbl/COTRTLIC.cbl
        lines 347 to 349 going forward and 363 to 365 going back, gated on a second
        edit flag with the same pass-through arrangement at lines 350 and 366 -- but
        the wildcards are supplied by the SERVICE and not by the caller: the baseline
        strings a percent sign before and after the trimmed input at lines 1155 to
        1163 before binding it. A value with no wildcard of its own therefore matches
        any description containing it, not only one equal to it. Send the text being
        searched for, not a pattern. This parameter takes ReferenceDescriptionFilter
        rather than the stored-description schema, because the baseline applies none
        of its maintenance-screen character rules to a filter.
      schema:
        $ref: '#/components/schemas/ReferenceDescriptionFilter'
...
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        Identity of the unit of work this request belongs to, propagated into the
        service log record and echoed on the response. Supply a value to tie this
        request to work already in progress upstream; omit it and the service
        mints one. A value that does not fit the published bound and character set
        is refused with 400 before any handler runs. A value whose shape is that of a
        primary account number -- only digits and the three separators, with
        thirteen or more digits -- is refused on the same terms, because this
        identity is echoed and logged. The refusal carries the shared problem body
        and the response correlation header, the same as any other 400, and it names
        the rule it applied without ever repeating the value it refused.
      schema:
        $ref: '#/components/schemas/CorrelationId'
  headers:
    XCorrelationId:
      description: >-
        Identity of the unit of work that produced this response. Echoed from the
        request where the caller supplied a conforming value, and minted by the
        service where it did not. Quote it when reporting a failure, because it is
        the route from this response to the operational record.
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'

    # WHY : Refactoring Rationale: this header is declared because the service
    #       actually sends it and an earlier revision of this document did not say
    #       so. The 401 is rendered by the shared handler
    #       common/error/ApiErrorSecurityHandlers.UnauthenticatedEntryPoint, which
    #       delegates to the framework's BearerTokenAuthenticationEntryPoint FIRST
    #       and only then writes the body, precisely so this challenge is composed
    #       exactly as the OAuth 2.0 bearer-token specification requires. A client's
    #       token-refresh logic reads this header, so leaving it undeclared left a
    #       generated client with no typed access to the one header it needs on the
    #       one response it needs it on.
    # WHY : Assumptions: the value is not enumerated and no pattern is written for
    #       it. The framework composes it from the error it detected -- a bare
    #       scheme when no token was presented, and a scheme with an error code and
    #       description when one was presented and rejected -- so pinning a shape
    #       here would either be wrong for one of those two cases or would have to
    #       be widened until it constrained nothing.
...
  responses:
    BadRequest:
      description: >-
        The request could not be accepted as supplied. This covers a malformed or
        out-of-range path or query value, a body that fails a declared constraint,
        a cursor that cannot be opened or was replayed against the wrong filter,
        collection or direction, a correlation header outside the published bound,
        and a write whose submitted values are identical to the ones already stored.
        The fieldErrors array names every offending field rather than only the first,
        for the reason recorded on the FieldError schema.


        The unchanged-submission refusal is the one member of that list that is a
        business rule rather than a syntactic check, and it belongs here rather than
        on 409 because it is decided from the request alone, before any row is
        contended: the baseline compares the submitted values with the ones it
        fetched and re-displays without writing. Its sentence is carried verbatim and
        is published on the ReferenceMessageCatalogue schema in both of the two
        wordings the baseline declares -- noChangeDetectedSingleRecord for a
        single-record replace and noChangeDetectedList for the list screen -- with a
        field entry naming the member that was resubmitted unchanged.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    # WHY : Refactoring Rationale: this response now states WHICH code and sentence
    #       the body carries, because an earlier revision promised the shared problem
    #       shape without naming its contents and the security chain did not produce
    #       one at all. A refusal decided by the filter chain never reaches a
    #       controller, so it never reaches the shared @RestControllerAdvice, and the
    #       framework default answered with a status and a challenge header and no
    #       body -- the exact opposite of what this response declared. The chain now
    #       installs common/error/ApiErrorSecurityHandlers, whose 401 renderer emits
    #       CARDDEMO-0401 with the sentence below.
    # WHY : Assumptions: the correlation identifier in the body is the SAME value as
    #       the header beside it and as the log line for this refusal, the renderer
    #       reading it from the diagnostic context the correlation filter established
    #       ahead of the security chain. That equality is the whole reason the header
    #       is required on a response whose body names no reason.
    Unauthorized:
      description: >-
        No usable bearer token was presented: the header was absent, the token was
        malformed, its signature or issuer did not validate, or it had expired.
        This is distinct from 403 and from 409, and the distinction is not
        cosmetic -- a caller acts on 401 by obtaining a token, on 403 by obtaining
        a different one, and on 409 by changing its request. No body member names
        the group the caller lacks, because at this point no verified identity
        exists to compare against one. The body carries code CARDDEMO-0401 and the
        sentence "Sign on to continue", identically on all four of those causes,
        alongside the bearer-token challenge below.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
        WWW-Authenticate:
          $ref: '#/components/headers/WwwAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: >-
        A valid token was presented, but it does not carry the authority this
        operation requires. In practice this is a token holding only carddemo-user
        sent to one of the administrative write operations, which
        SecurityConfig.java restricts by method at its lines 99 to 106. The
        identity is established and the request is well formed; only the authority
        is missing, which is why this is neither 401 nor 400. The body carries code
        CARDDEMO-0403 and the sentence "Not authorized for this operation" -- the
        same pair the shared advice publishes for a denial decided by a
        method-security check, so a caller cannot tell from the response where in
        the stack the authority was checked.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalServerError:
      description: >-
        The request could not be completed for a reason the service does not
        attribute to the caller. Nothing in the body names an internal detail: the
        correlation identifier is the route to the operational record, which is
        where diagnostic information lives. None of the outcomes the baseline
        distinguishes reaches this status -- the referential refusal, the
        concurrency conflict, the failed lock and the unmatched key are each
        reported with their own status above -- so a response here means an
        unclassified fault and not a business outcome.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    TransactionCategoryPage:
      type: object
      description: >-
        One page of transaction categories, ordered by type code and then by
        category code. The envelope carries the same four members as every other
        page in this document, and its cursors are sealed to this collection, so a
        cursor issued here cannot be replayed against the transaction-type list.
      required:
        - items
        - firstKey
        - lastKey
        - hasNext
      additionalProperties: false
      properties:
        items:
          type: array
          description: >-
            The categories of this page, ordered by type code and then category
            code. Empty, never null, when the set is exhausted or the filters
            matched nothing.
          items:
            $ref: '#/components/schemas/TransactionCategory'
        firstKey:
          oneOf:
            - $ref: '#/components/schemas/CursorToken'
            - type: 'null'
          description: >-
            The backward paging position over the composite key, or null when items
            is empty.
        lastKey:
          oneOf:
            - $ref: '#/components/schemas/CursorToken'
            - type: 'null'
          description: >-
            The forward paging position over the composite key, or null when items
            is empty.
        hasNext:
          type: boolean
          description: >-
            Whether a further page follows, settled by reading one row beyond the
            page.
      examples:
        - items:
            - typeCd: '01'
              catCd: '0001'
              description: Regular Sales Draft
              version: 1
          firstKey: v1.dHJhbnNhY3Rpb24tY2F0ZWdvcmllcw.REDACTED-EXAMPLE-SIGNATURE-not-a-real-mac-F
          lastKey: v1.dHJhbnNhY3Rpb24tY2F0ZWdvcmllcw.REDACTED-EXAMPLE-SIGNATURE-not-a-real-mac-L
          hasNext: true
...